Create new Groups with a consistent set of default properties.
This is part of a larger fix for #6450. Previously new route groups were
created with different default properties depending on which code created
the new group. The group tab method had all properties set and that is now
going to be the default for all other methods (mixer strip button, etc).